pyproject.toml
[tool.poetry]
name = "llm"
version = "0.1.0"
description = "bigly language model"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
llama-cpp-python = "^0.2.64"
fastapi = {extras = ["standard"], version = "^0.110.2"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
